Unify type inference by introducing a separate "high-level IR" (HIR)
Status quo: in the _ast2ir.py pass, we build an IR with untyped
operations: Call, Const and control flow ops (Loop, IfElse, etc.).
Then the typeinfer pass modifies this IR in-place to transform the
Call and Const calls using the @impl registration mechanism, and
populates type information for control flow ops using the infer_types()
method. This makes a few things awkward: the _ast2ir.py code is messy;
the typeinfer pass requires special hacks to handle control flow;
function call inlining requires non-trivial in-place mutation of the IR;
the IR has to accommodate "intermediate" ops and states to represent
the untyped program; it is difficult to undo the IR mutation done
by the typeinfer pass, in case we wanted to implement some sort of
iterative constant folding algorithm; etc.

With this change:
- We introduce a separate high-level intermediate representation (HIR),
  to which the AST is initially converted (ast2hir.py).
- Typeinfer pass is replaced with the HIR-to-IR pass (hir2ir.py).
- Type inference logic specific to control flow is no longer a special
  case: control flow operations are handled using the @impl registration
  system.
- Intermediate IR operations like Load, Store, Call, Const are removed.
- The typeinfer pass, TypingContext class and infer_types() method are
  removed.
- IfElse and Loop operations are simplified by removing IfElseResults/
  CarriedVariables objects. Loop's body variables are now modeled as
  block arguments.
- Every operation is now a TypedOperation (to be removed in a follow-up).

We also use this opportunity to improve error messages for type errors
caused by control flow. For example:

    @ct.kernel
    def foo(x):
        if ct.bid(0) == 0:
            a = x
        else:
            a = 1

Previously, this would trigger an error like this:

    Types mismatch for a.1 in propagation: Array[float32,(?,?):(?,1)] != int32
      In file "example.py", lines 6--9, col 5--22:
            if ct.bid(0) == 0:

With this change, this will not trigger any error because `a` is not
actually used. But if we add a use at the end, e.g. ct.printf("%s", a), the
new error message is like this:

    Type of `a` depends on path taken: int32 (line 9) vs. Array[float32,(?,?):(?,1)] (line 7)
      In file "example.py", line 9, col 9:
            a = 1

Type mismatches in loops are handled similarly. The logic for propagating types
through control flow is now unified in PhiState.propagate().

Signed-off-by: Greg Bonik <gbonik@nvidia.com>