Skip if-else results & loop variables of InvalidType
Say we have two branches of if-else that set the same local name
to two different types:

    if cond:
        x = 1
    else:
        x = ct.ones((3, 4), dtype=ct.float32)

This is not by itself an error unless `x` is used later.
In the current implementation, we would generate an IfElse op
with a result of an InvalidType, and the DCE pass would be responsible
for removing it, so that the bytecode generation wouldn't trip on it
later.

This works OK but 1) is somewhat fragile because of the reliance on the
DCE pass; and 2) makes it harder to make a change that removes aggregate
types like tuples from the IR.

In this commit, we remove the InvalidType variables from control flow
results and replace them with undefined variables. We achieve this by
delaying the setting of yielded results of Break/Continue/EndBranch ops.

Signed-off-by: Greg Bonik <gbonik@nvidia.com>